Move src/runner persistence service into src/lib/persistence
RunnerManager (plus runnerId, runnerIndex, runnerLoadError) is a
generic-shaped service — storage + migration orchestration — so it
belongs alongside the rest of src/lib rather than as its own top-level
directory. Its context (runnerManagerContext.tsx) already moved to
src/lib/contexts in the previous commit.
